Temporarily pin nixpkgs
This commit pins nixpkgs to the last known good
version for plover.

It seems like plover is doing some setuptools
magic to load its gui plugins, and that no longer
works with a newer version of setuptools.

Tried pinning setuptools to an older version,
but that doesn't seem to resolve the issue.

Pinning nixpkgs now so that the flake works for
now until there is further time to investigate,
or if someone comes up with a pull request.